refactor: rename plugin.ts to typedoc-plugin-external-module-name.ts for better debug experience
