+++ index.js
@@ -1,4 +1,4 @@
-var plugin = require('./plugin');
+var plugin = require('./typedoc-plugin-external-module-name');
 module.exports = function(PluginHost) {
   var app = PluginHost.owner;
   app.converter.addComponent('external-module-name', plugin.ExternalModuleNamePlugin);
